Refactor v3 manifest building into own class
